backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
click==8.3.0
colorama==0.4.6
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.121.2
greenlet==3.2.4
h11==0.16.0
httptools==0.7.1
idna==3.11
psycopg2-binary==2.9.11
pydantic==2.12.4
pydantic-core==2.41.5
pyjwt==2.8.0
pymssql==2.3.9
pymysql==1.1.2
python-dotenv==1.2.1
pyyaml==6.0.3
sniffio==1.3.1
sqlalchemy==2.0.23
starlette==0.49.3
typing-extensions==4.15.0
typing-inspection==0.4.2
uvicorn==0.27.1
watchfiles==1.1.1
websockets==15.0.1